backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==4.2.3
django-cors-headers==4.1.0
django-crum==0.7.9
djangorestframework==3.14.0
jedi==0.19.1
mysqlclient==2.1.1
packaging==24.1
parso==0.8.4
pudb==2024.1.2
Pygments==2.18.0
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urwid==2.6.15
urwid_readline==0.14
wcwidth==0.2.13
